+++ src/test/scala/stellar/sdk/util/ByteArraysSpec.scala
@@ -27,10 +27,10 @@ class ByteArraysSpec extends Specification with ArbitraryInput {
 
   "trimming a byte array" should {
     "remove trailing zeros" >> {
-      trimmedByteArray(Array()) mustEqual Array()
-      trimmedByteArray("hello".getBytes("UTF-8")) mustEqual "hello".getBytes("UTF-8")
-      trimmedByteArray("hello\u0000\u0000".getBytes("UTF-8")) mustEqual "hello".getBytes("UTF-8")
-      trimmedByteArray("hello\u0000there".getBytes("UTF-8")) mustEqual "hello\u0000there".getBytes("UTF-8")
+      trimmedByteArray(Array()) must beEmpty
+      trimmedByteArray("hello".getBytes("UTF-8")) mustEqual "hello".getBytes("UTF-8").toSeq
+      trimmedByteArray("hello\u0000\u0000".getBytes("UTF-8")) mustEqual "hello".getBytes("UTF-8").toSeq
+      trimmedByteArray("hello\u0000there".getBytes("UTF-8")) mustEqual "hello\u0000there".getBytes("UTF-8").toSeq
     }
   }